renderer: initial 3d rendering

+++ renderer/directx/directXShaderLoader.cpp
@@ -10,8 +10,7 @@
 #include <utilities/logging/logger.hpp>
 #include <utilities/wildcardMatch.hpp>
 
-const std::map<ShaderType, std::string> compilers
-{
+const std::map<ShaderType, std::string> compilers {
 	{ShaderType::Compute, "cs_5_0"},
 	{ShaderType::Domain, "ds_5_0"},
 	{ShaderType::Geometry, "gs_5_0"},
@@ -20,21 +19,21 @@ const std::map<ShaderType, std::string> compilers
 	{ShaderType::Vertex, "vs_5_0"},
 };
 
-bool compileShaderFromFile(const std::string& filename,	const D3D_SHADER_MACRO* defines,
+bool compileShaderFromFile(const std::string& filename, const D3D_SHADER_MACRO* defines,
 	LPCSTR entry_point, std::string shader_compiler, ID3DBlob** bytecode)
 {
 	UINT compiler_flags = D3DCOMPILE_ENABLE_STRICTNESS | D3DCOMPILE_PACK_MATRIX_COLUMN_MAJOR;
 
-	#ifdef _DEBUG
-		compiler_flags |= D3DCOMPILE_DEBUG;
-	#endif
+#ifdef _DEBUG
+	compiler_flags |= D3DCOMPILE_DEBUG;
+#endif
 
 	ID3DBlob* error_blob = nullptr;
 	auto utf16_filename = Encoding::toWindowsString(filename);
 
-	HRESULT hr = D3DCompileFromFile(utf16_filename.c_str(), defines,
-		D3D_COMPILE_STANDARD_FILE_INCLUDE, entry_point, shader_compiler.c_str(),
-		compiler_flags, 0, bytecode, &error_blob);
+	HRESULT hr =
+		D3DCompileFromFile(utf16_filename.c_str(), defines, D3D_COMPILE_STANDARD_FILE_INCLUDE,
+			entry_point, shader_compiler.c_str(), compiler_flags, 0, bytecode, &error_blob);
 
 	if (FAILED(hr))
 	{
@@ -50,28 +49,29 @@ bool compileShaderFromFile(const std::string& filename,	const D3D_SHADER_MACRO* 
 	return SUCCEEDED(hr) == TRUE;
 }
 
-bool compileShaderFromBuffer(const std::string& filename, uint8_t* buffer,
-	const unsigned int size, const D3D_SHADER_MACRO* defines,
-	LPCSTR entry_point, std::string shader_compiler, ID3DBlob** bytecode)
+bool compileShaderFromBuffer(const std::string& filename, uint8_t* buffer, const uint64_t size,
+	const D3D_SHADER_MACRO* defines, LPCSTR entry_point, std::string shader_compiler,
+	ID3DBlob** bytecode)
 {
 	UINT compiler_flags = D3DCOMPILE_ENABLE_STRICTNESS | D3DCOMPILE_PACK_MATRIX_COLUMN_MAJOR;
 
-	#ifdef _DEBUG
-		compiler_flags |= D3DCOMPILE_DEBUG;
-	#endif
+#ifdef _DEBUG
+	compiler_flags |= D3DCOMPILE_DEBUG;
+#endif
 
 	ID3DBlob* error_blob = nullptr;
 
-	HRESULT hr = D3DCompile(buffer, size, filename.c_str(), defines,
-		D3D_COMPILE_STANDARD_FILE_INCLUDE, entry_point, shader_compiler.c_str(),
-		compiler_flags, 0, bytecode, &error_blob);
+	HRESULT hr =
+		D3DCompile(buffer, size, filename.c_str(), defines, D3D_COMPILE_STANDARD_FILE_INCLUDE,
+			entry_point, shader_compiler.c_str(), compiler_flags, 0, bytecode, &error_blob);
 
 	if (FAILED(hr))
 	{
 		if (error_blob != nullptr && error_blob->GetBufferSize() > 0)
 		{
-			logWarning("failed to compile shader " + std::string((char*)error_blob->GetBufferPointer(), error_blob->GetBufferSize()));
-            error_blob->Release();
+			logWarning("failed to compile shader " +
+				std::string((char*)error_blob->GetBufferPointer(), error_blob->GetBufferSize()));
+			error_blob->Release();
 			// log error message
 		}
 	}
@@ -92,7 +92,7 @@ bool loadShaderFromFile(const std::string& filename, ID3DBlob** bytecode)
 	return bytecode != nullptr;
 }
 
-bool loadShaderFromBuffer(uint8_t* buffer, const unsigned int size, ID3DBlob** bytecode)
+bool loadShaderFromBuffer(uint8_t* buffer, const uint64_t size, ID3DBlob** bytecode)
 {
 	if (FAILED(D3DCreateBlob(size, bytecode)))
 	{
@@ -107,8 +107,8 @@ bool loadShaderFromBuffer(uint8_t* buffer, const unsigned int size, ID3DBlob** b
 	return bytecode != nullptr;
 }
 
-bool loadShaderBytecode(const std::string& filename, uint8_t* buffer,
-	const unsigned int size, ID3DBlob** bytecode, const ShaderType type)
+bool loadShaderBytecode(const std::string& filename, uint8_t* buffer, const uint64_t size,
+	ID3DBlob** bytecode, const ShaderType type)
 {
 	if (wildcardMatch("*.cso", filename.c_str()))
 	{
@@ -143,7 +143,8 @@ bool loadShaderBytecode(const std::string& filename, uint8_t* buffer,
 		}
 		else
 		{
-			if (!compileShaderFromBuffer(filename, buffer, size, nullptr, "main", compiler, bytecode))
+			if (!compileShaderFromBuffer(filename, buffer, size, nullptr, "main", compiler,
+					bytecode))
 			{
 				// log error
 				return false;
