+++ renderer/directx/directXShaderLoader.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+
 #include <d3d11.h>
 
 enum class ShaderType
@@ -13,5 +14,5 @@ enum class ShaderType
 	Hull
 };
 
-bool loadShaderBytecode(const std::string& filename, uint8_t* buffer,
-	const unsigned int size, ID3DBlob** bytecode, const ShaderType type);
+bool loadShaderBytecode(const std::string& filename, uint8_t* buffer, const uint64_t size,
+	ID3DBlob** bytecode, const ShaderType type);
